Print tail of log file and message upon successful app start-up.

+++ scripts/install-app.sh
@@ -18,6 +18,7 @@
 
 APP_VARS="APP_DIR APP_GID APP_LOGFILE APP_NAME APP_PIDFILE APP_PORT \
 APP_RUNDIR APP_UID APP_VARDIR"
+POLL_COUNT=10
 POLL_INTERVAL=2
 
 abort() {
@@ -205,19 +206,24 @@ for dryrun in true false; do
     install_app_and_config
 done
 
-if signal_app HUP; then
-    sleep $SLEEP_INTERVAL "Waiting for app to restart"
-elif [ "$distro_name" = ubuntu ]; then
+
+if signal_app HUP && [ "$distro_name" = ubuntu ]; then
     /bin/rm -f $APP_PIDFILE
     service uwsgi restart
     sleep $POLL_INTERVAL "Waiting for app to start"
     i=0
 
-    until [ -e $APP_PIDFILE -o $i -ge 10 ]; do
+    until [ -e $APP_PIDFILE -o $i -ge $POLL_COUNT ]; do
 	sleep $POLL_INTERVAL "Waiting for app to start"
 	i=$((i + 1))
     done
+else
+    sleep $KILL_INTERVAL "Waiting for app to start"
 fi
 
-tail_log_file
-
+if [ -e $APP_PIDFILE ]; then
+    tail_log_file
+    printf "%s\n" "App started successfully"
+else
+    abort "%s\n" "App may not have started"
+fi
